app-requirements.txt
antlr4-python3-runtime==4.8
Click==7.0
Cython==0.29.17
Flask==1.1.1
Flask-Cors==3.0.8
hdbscan==0.8.26
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
llvmlite==0.32.0
MarkupSafe==1.1.1
numba==0.49.0
numpy==1.18.1
pandas==1.0.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
tbb==2019.0
umap-learn==0.4.2
Werkzeug==0.16.1
